ROUTER
Include in showcase url / product before url coming by json
I put the showcase inside the route /
I created a route to the product page that receives the url / product
and a paratrooper slug that is the url of the product.

I changed the summer node-sass to the most current one, because it was
giving an error when running the project in windows.